Major re-ordering to make non-modal
The use of modal dialog boxes and uiwait meant that only user input could progress the process, but to unit test we want to be able to use callbacks to simulate user input. This commit re-orders things to make it work but should not break things, but that remains to be unit-tested.