Fix legacy startup
The problem was that the 'LocalDir' param is used during testing because otherwise the Local setup is assumed to be in the same parent dir as Bpod_Gen2. The slightly different pathing logic resulted in an attempt to access a non-existent BpodObject.Path field, so now the object is initialised with this "assumed" at startup so any checks when completing startup can refer to it. Not pretty, but it works.